feat(sv): add `scrollerRef` prop and import cjs version of gsap and ScrollTrigger
1. `scrollerRef`:
If we don't pass `scroller` prop to ScrollTrigger instance,
and then `window` will be thing whose scrollbar is linked to the
ScrollTrigger.

`scrollerRef` is the element to replace `window` to be associated
with ScrollTrigger. But, if `scrollerRef` is `undefined`, and `window`
will still be the fallback.

2. import cjs version of gsap
For Keystone CMS (`packages/cms`), there will be a webpack error due to
importing `@story-telling-reporter/react-scrollable-video`.

The root cause is webpack cannot handle ESModule well since
we do not specify `"type": "module"` in the `packages/cms/package.json`.

Therefore, we work around this error by importing cjs version of gsap.