fix(sta): muted button still does not disapear even we add cleanup function in useEffect
The bug description details is in this [asana task]
(https://app.asana.com/0/1205812314629278/1207546924043990/f).

When client side routing to another webpage, the cleanup function in the
`useEffect` is not executed. Therefore, the muted button could not be removed
from document.body.

Currently, I cannot come up with a good solution to address this bug.
Hence, I will address this bug by updating `ScrollableVideo` component
instead.

The way to fix the bug is to remove containing blocks in the
`ScrollableVideo` component.

If there is no containing blocks wrapping `ScrollToAudio`, and then
`FixedMutedButton` will be fixed in the viewport as we expect rather than
fixed in the containing blocks.